# api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: Fullstack Workspace API
  version: 1.0.0
servers:
  - url: /api
    description: API base path
paths:
...
  /auth/email:
    post:
      summary: 发送邮件验证码
      operationId: sendEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEmail"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
  /auth/register:
    post:
      summary: 用户注册
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
  /auth/login:
    post:
      summary: 用户登录
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultToken"
  /auth/refresh:
    post:
      summary: 刷新 Token
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultToken"
  /auth/logout:
    post:
      summary: 用户登出
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
  /auth/change-password:
    post:
      summary: 修改密码
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
  /auth/forgot-password:
    post:
      summary: 忘记密码-发送验证码
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
  /auth/reset-password:
    post:
      summary: 忘记密码-重置
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultVoid"
  /auth/me:
    get:
      summary: 当前用户信息
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResultUserMe"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResult:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: "0 成功, 1000 参数校验错误"
        message:
          type: string
        data:
          type: object
    ApiResultVoid:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              type: object
              nullable: true
...
    ApiResultToken:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              $ref: "#/components/schemas/TokenResponse"
...
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    LogoutRequest:
      type: object
      properties:
        refreshToken:
          type: string
        deviceId:
          type: string
    ChangePasswordRequest:
      type: object
      required:
        - newPassword
      properties:
        oldPassword:
          type: string
        emailCode:
          type: string
        newPassword:
          type: string
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required:
        - email
        - emailCode
        - newPassword
      properties:
        email:
          type: string
          format: email
        emailCode:
          type: string
        newPassword:
          type: string
    UserMe:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
        gender:
          type: integer
        birthday:
          type: string
          format: date
        bio:
          type: string
...
    SendEmail:
      type: object
      required:
        - email
        - usage
      properties:
        email:
          type: string
          format: email
        usage:
          type: string
    UserRegister:
      type: object
      required:
        - email
        - password
        - emailCode
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        emailCode:
          type: string
    UserLogin:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        loginType:
          type: string
        emailCode:
          type: string
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
        deviceId:
          type: string
    ApiResultUserMe:
      allOf:
        - $ref: "#/components/schemas/ApiResult"
        - properties:
            data:
              $ref: "#/components/schemas/UserMe"
...